Refactor notifier tests
The tests for `TaskNotifier` and `RequestNotifier` were simply duplicates of each other. Plus, the functionality is actually in the `BaseNotifier`. All the child notifiers really do is implement the formatting methods.

So, this final refactor removes the duplication in the tests and has them focus on what each class is doing.
